# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: PrestaRapi - Sistema de Gestión de Préstamos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/recover-request:
    post:
      operationId: requestPasswordRecovery
      tags: [auth]
      summary: Request password recovery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryRequest"
      responses:
        "200":
          description: Recovery request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/recover-approve/{userId}:
    post:
      operationId: approvePasswordRecovery
      tags: [auth]
      summary: Approve password recovery and set new password
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRecoveryRequest"
      responses:
        "200":
          description: Password set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApproveRecoveryResponse"
  /agencies:
    get:
      operationId: getAgencies
      tags: [agencies]
      summary: List agencies
      responses:
        "200":
          description: List of agencies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agency"
    post:
      operationId: createAgency
      tags: [agencies]
      summary: Create agency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgencyRequest"
      responses:
        "201":
          description: Agency created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
  /agencies/{id}:
    get:
      operationId: getAgency
      tags: [agencies]
      summary: Get agency
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
    put:
      operationId: updateAgency
      tags: [agencies]
      summary: Update agency
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgencyRequest"
      responses:
        "200":
          description: Agency updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
    delete:
      operationId: deleteAgency
      tags: [agencies]
      summary: Soft-delete agency
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: List clients
      parameters:
        - name: agencyId
          in: query
          schema:
            type: integer
        - name: assignedTo
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client with account statement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client with full account statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientWithStatement"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Soft-delete client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /clients/{id}/reassign:
    post:
      operationId: reassignClient
      tags: [clients]
      summary: Reassign client to another user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReassignRequest"
      responses:
        "200":
          description: Client reassigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
    RecoveryRequest:
      type: object
      properties:
        username:
          type: string
      required:
        - username
    ApproveRecoveryRequest:
      type: object
      properties:
        newPassword:
          type: string
      required:
        - newPassword
    ApproveRecoveryResponse:
      type: object
      properties:
        newPassword:
          type: string
        userName:
          type: string
      required:
        - newPassword
        - userName
    Agency:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        active:
          type: boolean
        currency:
          type: string
        baseRate:
          type: number
          description: Default interest rate (e.g. 0.10 = 10%)
        graceDaysDisbursement:
          type: integer
          description: Grace days after disbursement before interest accrues
        graceDaysPayment:
          type: integer
          description: Grace days after payment before new interest accrues
        paymentMethods:
          type: array
          items:
            type: string
        banks:
          type: array
          items:
            $ref: "#/components/schemas/BankConfig"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - city
        - active
        - currency
        - baseRate
        - graceDaysDisbursement
        - graceDaysPayment
        - paymentMethods
        - banks
        - createdAt
    BankConfig:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
      required:
        - name
        - url
    CreateAgencyRequest:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        currency:
          type: string
        baseRate:
          type: number
        graceDaysDisbursement:
          type: integer
        graceDaysPayment:
          type: integer
        paymentMethods:
          type: array
          items:
            type: string
        banks:
          type: array
          items:
            $ref: "#/components/schemas/BankConfig"
      required:
        - name
        - city
        - currency
        - baseRate
        - graceDaysDisbursement
        - graceDaysPayment
        - paymentMethods
        - banks
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [superadmin, admin, manager, promoter]
        agencyIds:
          type: array
          items:
            type: integer
        blocked:
          type: boolean
        avatar:
          type: string
        needsPasswordRecovery:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - username
        - role
        - agencyIds
        - blocked
        - needsPasswordRecovery
        - createdAt
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        phone2:
          type: string
        agencyId:
          type: integer
        agencyName:
          type: string
        assignedToId:
          type: integer
        assignedToName:
          type: string
        assignedToRole:
          type: string
        active:
          type: boolean
        capitalBalance:
          type: number
        currentInterest:
          type: number
        overdueInterest:
          type: number
        overdueQuincenas:
          type: integer
        createdAt:
          type: string
          format: date
      required:
        - id
        - name
        - agencyId
        - assignedToId
        - active
        - capitalBalance
        - currentInterest
        - overdueInterest
        - overdueQuincenas
        - createdAt
    ClientWithStatement:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            loans:
              type: array
              items:
                $ref: "#/components/schemas/Loan"
            movements:
              type: array
              items:
                $ref: "#/components/schemas/Movement"
    CreateClientRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        phone2:
          type: string
        agencyId:
          type: integer
        assignedToId:
          type: integer
      required:
        - name
        - agencyId
        - assignedToId
    ReassignRequest:
      type: object
      properties:
        assignedToId:
          type: integer
      required:
        - assignedToId
    Loan:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        agencyId:
          type: integer
        agencyName:
          type: string
        promoterId:
          type: integer
        promoterName:
          type: string
        originalCapital:
          type: number
        currentCapital:
          type: number
        rate:
          type: number
        periodicity:
          type: string
          enum: [daily, weekly, biweekly, monthly]
        status:
          type: string
          enum: [pending_approval, approved, pending_disbursement, active, rejected, settled]
        overdueInterest:
          type: number
        currentInterest:
          type: number
        disbursementDate:
          type: string
          format: date
        nextCutDate:
          type: string
          format: date
        disbursementMethod:
          type: string
        disbursementProof:
          type: string
        inDefault:
          type: boolean
        rejectReason:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - agencyId
        - originalCapital
        - currentCapital
        - rate
        - periodicity
        - status
        - overdueInterest
        - currentInterest
        - inDefault
        - createdAt
...
    Movement:
      type: object
      properties:
        id:
          type: integer
        loanId:
          type: integer
        type:
          type: string
          enum: [disbursement, interest_accrual, payment_overdue_interest, payment_current_interest, payment_capital, forgiveness, reversal]
        amount:
          type: number
        concept:
          type: string
        date:
          type: string
          format: date
        userId:
          type: integer
        userName:
          type: string
        reversed:
          type: boolean
        image:
          type: string
      required:
        - id
        - loanId
        - type
        - amount
        - concept
        - date
        - reversed