# lib/api-zod/src/generated/types/createAgencyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrestaRapi - Sistema de Gestión de Préstamos
 * OpenAPI spec version: 1.0.0
 */
import type { BankConfig } from "./bankConfig";

export interface CreateAgencyRequest {
  name: string;
  city: string;
  currency: string;
  baseRate: number;
  graceDaysDisbursement: number;
  graceDaysPayment: number;
  paymentMethods: string[];
  banks: BankConfig[];
}